+++ src/app/pages/types/trailer-item.ts
@@ -1,0 +1,4 @@
+import { MovieWithTrailer } from '../models/movie-with-trailer';
+import { TVWithTrailer } from '../models/tv-with-trailer';
+
+export type TrailerItem = MovieWithTrailer | TVWithTrailer;